+++ src/Git/GitIntegration.php
@@ -2,14 +2,14 @@
 
 namespace Becklyn\Hosting\Git;
 
+use Symfony\Component\Process\Process;
+
 /**
  * Integration with git.
  */
 class GitIntegration
 {
-    /**
-     * @var string
-     */
+    /** @var string */
     private $projectDir;
 
 
@@ -30,10 +30,10 @@ class GitIntegration
             return null;
         }
 
-        $git = $this->run('command -v git');
+        $git = $this->run(["command", "-v", "git"]);
 
         return null !== $git
-            ? $this->run("{$git} rev-parse HEAD")
+            ? $this->run([$git, "rev-parse", "HEAD"])
             : null;
     }
 
@@ -41,9 +41,12 @@ class GitIntegration
     /**
      * Runs the given command.
      */
-    private function run (string $command) : ?string
+    private function run (array $command) : ?string
     {
-        $result = \trim(\shell_exec($command) ?? "");
+        $process = new Process($command);
+        $process->mustRun();
+
+        $result = \trim($process->getOutput());
 
         return "" !== $result
             ? $result