+++ src/Listener/MonitoringTokenListener.php
@@ -4,14 +4,12 @@ namespace Becklyn\Hosting\Listener;
 
 use Becklyn\Hosting\Config\HostingConfig;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
-use Symfony\Component\HttpKernel\Event\FilterResponseEvent;
+use Symfony\Component\HttpKernel\Event\ResponseEvent;
 use Symfony\Component\HttpKernel\KernelEvents;
 
 class MonitoringTokenListener implements EventSubscriberInterface
 {
-    /**
-     * @var string
-     */
+    /** @var string */
     private $uptimeHtmlEmbed;
 
 
@@ -25,7 +23,7 @@ class MonitoringTokenListener implements EventSubscriberInterface
 
     /**
      */
-    public function onResponse (FilterResponseEvent $event) : void
+    public function onResponse (ResponseEvent $event) : void
     {
         // skip if not master request
         if (!$event->isMasterRequest() || null === $this->uptimeHtmlEmbed)
@@ -37,7 +35,7 @@ class MonitoringTokenListener implements EventSubscriberInterface
         $contentType = $response->headers->get("Content-Type");
 
         // skip if not HTML response
-        if (null === $contentType || false === \mb_strpos($contentType, "text/html"))
+        if (null === $contentType || false === \strpos($contentType, "text/html"))
         {
             return;
         }