+++ src/Project/ProjectVersion.php
@@ -3,8 +3,7 @@
 namespace Becklyn\Hosting\Project;
 
 use Becklyn\Hosting\Git\GitIntegration;
-use Psr\Cache\CacheItemInterface;
-use Psr\Cache\CacheItemPoolInterface;
+use Symfony\Contracts\Cache\CacheInterface;
 
 /**
  * Fetches the version of the project.
@@ -13,52 +12,27 @@ class ProjectVersion
 {
     const CACHE_KEY = "becklyn.hosting.version";
 
-
-    /**
-     * @var string|null
-     */
+    /** @var string|null */
     private $version;
 
-
-    /**
-     * @var bool
-     */
+    /** @var bool */
     private $initialized = false;
 
-
-    /**
-     * @var GitIntegration
-     */
+    /** @var GitIntegration */
     private $gitIntegration;
 
-
-    /**
-     * @var CacheItemPoolInterface
-     */
-    private $cachePool;
-
-
-    /**
-     * @var CacheItemInterface
-     */
-    private $cacheItem;
+    /** @var CacheInterface */
+    private $cache;
 
 
     /**
      */
-    public function __construct (GitIntegration $gitIntegration, CacheItemPoolInterface $cachePool)
+    public function __construct (GitIntegration $gitIntegration, CacheInterface $cache)
     {
-        $this->cachePool = $cachePool;
-        $this->cacheItem = $cachePool->getItem(self::CACHE_KEY);
         $this->gitIntegration = $gitIntegration;
-
-        // fetch data from cache
-        if ($this->cacheItem->isHit() && null !== $this->cacheItem->get())
-        {
-            $this->version = $this->cacheItem->get();
-            $this->initialized = true;
-        }
+        $this->cache = $cache;
     }
+
 
     /**
      */
@@ -66,11 +40,12 @@ class ProjectVersion
     {
         if (!$this->initialized)
         {
-            $this->version = $this->gitIntegration->fetchHeadCommitHash();
-            $this->initialized = true;
+            $this->version = $this->cache->get(
+                self::CACHE_KEY,
+                [$this->gitIntegration, "fetchHeadCommitHash"]
+            );
 
-            $this->cacheItem->set($this->version);
-            $this->cachePool->save($this->cacheItem);
+            $this->initialized = true;
         }
 
         return $this->version;